02 - Agent with tools/pyproject.toml
[project]
name = "basic-ai-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
"langchain",
"wikipedia",
"langchain-community",
"langchain-core",
"langchain-openai",
"python-dotenv",
"pydantic",
"duckduckgo-search",
"langchain-google-genai",
"ddgs"
]